pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Based on Moral Machine Dataset, this project aims to train a model to predict human moral decisions in various scenarios."
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "numpy>=2.3.5",
    "onnxscript>=0.6.0",
    "pandas>=3.0.0",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ipython>=9.10.0",
    "jupyter>=1.1.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"
